authentication added

+++ admin_portal/app.py
@@ -16,11 +16,15 @@ app.register_blueprint(site)
 
 @app.route('/')
 def homepage():
-    return redirect(url_for('login'))
+    if sess['authenticated'] is not True:
+        return redirect(url_for('login'))
+    return redirect(url_for(dashboard))
 
 
 @app.route('/check-credentials/', methods=["POST"])
 def check_credentials():
+    if sess['authenticated'] is not True:
+        return redirect(url_for('login'))
     error = None
     if request.method == 'POST':
         if request.form.get('username', None) != 'admin' or request.form.get('password', None) != 'admin':
@@ -34,21 +38,30 @@ def check_credentials():
 
 @app.route('/login/')
 def login():
+    sess['authenticated'] = True
     return render_template('login.html')
 
 
 @app.route('/dashboard/')
 def dashboard():
+    if sess['authenticated'] is not True:
+        return redirect(url_for('login'))
+
     return render_template("dashboard.html")
 
 
 @app.route('/add-book/')
 def add():
+    if sess['authenticated'] is not True:
+        return redirect(url_for('login'))
     return render_template("add-book.html")
 
 
 @app.route('/add-book-verification/', methods=["POST", "GET"])
 def add_book_verification():
+    if sess['authenticated'] is not True:
+        return redirect(url_for('login'))
+
     if request.method == 'POST':
         try:
             title = request.form.get('title')
@@ -77,11 +90,17 @@ def add_book_verification():
 
 @app.route('/search-book/')
 def search_book():
+    if sess['authenticated'] is not True:
+        return redirect(url_for('login'))
+
     return render_template("search-book.html")
 
 
 @app.route('/search/', methods=["POST", "GET"])
 def search():
+    if sess['authenticated'] is not True:
+        return redirect(url_for('login'))
+
     if request.method == 'POST':
         query = request.form.get('query')
         result = Book.query.with_entities(Book.id, Book.title, Book.isbn, Book.author, Book.published_year,
@@ -98,17 +117,33 @@ def search():
 
 @app.route('/list-books/')
 def list_books():
+    if sess['authenticated'] is not True:
+        return redirect(url_for('login'))
+
     return render_template("list-books.html")
+
+
+@app.route('/logout/')
+def logout():
+    # end the session authentication
+    session.pop('authenticated', None)
+
+    return redirect(url_for("/"))
 
 
 @app.route('/delete-book/<id>', methods=["POST"])
 def delete_book(id):
+
+    if sess['authenticated'] is not True:
+        return redirect(url_for('login'))
+
+    # remove triangular brackets that comes with the request
     id = id[1:-1]
-    
+
     if request.method == 'POST':
         Book.query.filter(Book.id == id).delete()
         db.session.commit()
-    return redirect(url_for("dashboard"))
+
 
 @app.errorhandler(404)
 def page_not_found():
